+++ src/main/java/company/board_project/domain/match/match/dto/MatchEndDto.java
@@ -8,14 +8,14 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @Getter
 public class MatchEndDto {
-    private Long matchId;
-    private Long matchListId;
-    private Long homeTeamUserId;
-    private Long awayTeamUserId;
-    private Long homeTeamId;
-    private Long awayTeamId;
-    private Long homeTeamScore;
-    private Long awayTeamScore;
+    private long matchId;
+    private long matchListId;
+    private long homeTeamUserId;
+    private long awayTeamUserId;
+    private long homeTeamId;
+    private long awayTeamId;
+    private int homeTeamScore;
+    private int awayTeamScore;
     private String matchStatus;
 
 }
